cmd/geth: tidied up the source
